+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## main
 
+
 ### Features
 
 - `[jest-runtime]` Reduce redundant ReferenceError messages
@@ -7,6 +8,7 @@
 
 ### Fixes
 
+- `[jest-snapshot]` Fix mangled inline snapshot updates when used with Prettier 3 and CRLF line endings
 - `[jest-runtime]` Importing from `@jest/globals` in more than one file no longer breaks relative paths ([#15772](https://github.com/jestjs/jest/issues/15772))
 
 # Chore